migrate from uint32 to int
instead of just converting the ints to uint32 and back, we could use
ints directly

also create a new helpers file which is used by main.go
