+++ stellar/utils/utils.go
@@ -5,7 +5,6 @@ import (
 	//"log"
 	"bufio"
 	"os"
-	"encoding/binary"
 	"encoding/hex"
 	"fmt"
 	"net/http"
@@ -69,7 +68,7 @@ func ScanForStringWithCheckF() (string, error) {
 		return "", fmt.Errorf("Couldn't read user input")
 	}
 	inputString := scanner.Text()
-	if StringToFloat(inputString) == 0 {
+	if StoF(inputString) == 0 {
 		fmt.Println("Amount entered is not a float, quitting")
 		return "", fmt.Errorf("Amount entered is not a float, quitting")
 	}
@@ -97,52 +96,40 @@ func ScanRawPassword() (string, error) {
 	return password, nil
 }
 
-func StoI(a string) int {
-	temp, _ := strconv.Atoi(a)
-	// ignore error sicne we assume that we'll call this in the right place
-	return temp
-}
-
 func Timestamp() string {
 	return time.Now().Format(time.RFC850)
 }
 
-func Uint32toB(a uint32) []byte {
+func ItoB(a int) []byte {
 	// need to convert int to a byte array for indexing
-	temp := make([]byte, 4)
-	binary.LittleEndian.PutUint32(temp, a)
-	return temp
+	string1 := strconv.Itoa(a)
+	return []byte(string1)
 }
 
-func Uint32toS(a uint32) string {
-	// convert uint32 to int and then int to string
-	aInt := int(a)
-	aStr := strconv.Itoa(aInt)
+func ItoS(a int) string {
+	aStr := strconv.Itoa(a)
 	return aStr
 }
 
-func StoUint32(a string) uint32 {
-	// convert string to int
-	aInt, _ := strconv.Atoi(a)
-	return uint32(aInt)
+func BToI(a []byte) int {
+	x, _ := strconv.Atoi(string(a))
+	return x
 }
 
-func BToUint32(a []byte) uint32 {
-	return binary.LittleEndian.Uint32(a)
-}
-
-func FloatToString(a float64) string {
+func FtoS(a float64) string {
 	return fmt.Sprintf("%f", a)
 }
 
-func StringToFloat(a string) float64 {
+func StoF(a string) float64 {
 	x, _ := strconv.ParseFloat(a, 32)
 	// ignore this error since we hopefully call this in the right place
 	return x
 }
 
-func IntToString(a int) string {
-	return strconv.Itoa(a)
+func StoI(a string) int {
+	// convert string to int
+	aInt, _ := strconv.Atoi(a)
+	return aInt
 }
 
 func SHA3hash(inputString string) string {
